Move patched_conics to threebody
