+++ src/poliastro/threebody/patched_conics.py
